+++ modules/python_engine/tests/test_pyrunfile_1.py
@@ -1,0 +1,2 @@
+greeting = "hello Nelson"
+print(greeting)